schbench: move decision to stop into main thread
The message threads were checking gtod to see if they had exceeded the
runtime, which meant they had to use timeouts for fpost().  This changes
the main thread to sleep until the run time is up, and then flip the
global stopping var.

This allows the mesage threads to avoid sending a timeout to fpost,
lowering overhead.  The timeout can also hide synchronization bugs in
the post/wait setup, this way we'll find any problems more easily.

Signed-off-by: Chris Mason <clm@fb.com>