Add RPS mode
This adds -R (--rps) to try and simulate a server thread handing out
requests to workers.  Instead of focusing on wakeup latency, this
calculates the latency from the time the request is queued up for a
worker to the time the worker finishes the request.

To help normalize the results, we subtract the cputime (--cputime) of
each request from the latency number, so we end up with the extra time
we spent waiting between when the request was queued to the time we
completed it.

To normalize things a little more, a p99 summary is given where we
divide this extra latency by the cputime.  For example, with the default
cputime of 30ms, if the delay is 15ms, we'd print a latency of 50% of
the cputime.

The real advantage for -R is to use it along with --auto to process a range
of workloads all at once.

schbench -m 2 -t 23 -R 1 -a -r 10

Using -R 1 tells it to guess a good starting point, and keep bumping the
RPS until the delay gets too long.

Signed-off-by: Chris Mason <clm@fb.com>